Split into platform specific dlls
