+++ src/BugsnagUnity/BugsnagUnity.Android.csproj
@@ -7,6 +7,8 @@
     <Version></Version>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Remove="Native/**/*.cs" />
+    <Compile Include="Native/Android/**/*.cs" />
     <Reference Include="UnityEngine">
       <HintPath>$(UnityDir)\UnityEngine.dll</HintPath>
       <Private>false</Private>
