+++ src/BugsnagUnity/BugsnagUnity.Windows.csproj
@@ -7,6 +7,10 @@
     <Version></Version>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Remove="Native/**/*.cs" />
+    <Compile Include="Native/Windows/**/*.cs" />
+    <Compile Include="Native/Fallback/Configuration.cs" />
+    <Compile Include="Native/Fallback/Breadcrumbs.cs" />
     <Reference Include="UnityEngine">
       <HintPath>$(UnityDir)\UnityEngine.dll</HintPath>
       <Private>false</Private>
