+++ src/BugsnagUnity/BugsnagUnity.csproj
@@ -7,6 +7,7 @@
     <Version></Version>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Remove="Native/**/*.cs"/>
     <Reference Include="UnityEngine">
       <HintPath>$(UnityDir)\UnityEngine.dll</HintPath>
       <Private>false</Private>
