+++ src/BugsnagUnity/Client.cs
