--- src/BugsnagUnity/Clients/CocoaClient.cs
@@ -1,213 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Runtime.InteropServices;
-using AOT;
-using BugsnagUnity.Payload;
-
-namespace BugsnagUnity
-{
-  class CocoaClient : INativeClient
-  {
-    internal const string MacOsImport = "bugsnag-osx";
-    internal const string iOSImport = "__Internal";
-  
-    public IConfiguration Configuration { get; }
-
-    public IBreadcrumbs Breadcrumbs { get; }
-
-    public IDelivery Delivery { get; }
-
-    IntPtr NativeConfiguration { get; }
-
-    ICocoaWrapper Wrapper { get; }
-    
-    CocoaClient(IConfiguration configuration, IntPtr nativeConfiguration, ICocoaWrapper wrapper, IBreadcrumbs breadcrumbs)
-    {
-      Configuration = configuration;
-      NativeConfiguration = nativeConfiguration;
-      Wrapper = wrapper;
-
-      Wrapper.StartBugsnagWithConfiguration(NativeConfiguration);
-
-      Delivery = new Delivery();
-      Breadcrumbs = breadcrumbs;
-    }
-
-    public CocoaClient(MacOSConfiguration configuration) : this(configuration, configuration.NativeConfiguration, new MacOsWrapper(), new CocoaBreadcrumbs(configuration))
-    {
-    }
-
-    public CocoaClient(iOSConfiguration configuration) : this(configuration, configuration.NativeConfiguration, new iOSWrapper(), new CocoaBreadcrumbs(configuration))
-    {
-    }
-
-    public void PopulateApp(App app)
-    {
-      GCHandle handle = GCHandle.Alloc(app);
-
-      try
-      {
-        Wrapper.RetrieveAppData(GCHandle.ToIntPtr(handle), PopulateAppData);
-      }
-      finally
-      {
-        if (handle != null)
-        {
-          handle.Free();
-        }
-      }
-    }
-
-    [MonoPInvokeCallback(typeof(Action<IntPtr, string, string>))]
-    static void PopulateAppData(IntPtr instance, string key, string value)
-    {
-      var handle = GCHandle.FromIntPtr(instance);
-      if (handle.Target is App app)
-      {
-        app.AddToPayload(key, value);
-      }
-    }
-
-    public void PopulateDevice(Device device)
-    {
-      var handle = GCHandle.Alloc(device);
-
-      try
-      {
-        Wrapper.RetrieveDeviceData(GCHandle.ToIntPtr(handle), PopulateDeviceData);
-      }
-      finally
-      {
-        handle.Free();
-      }
-    }
-
-    [MonoPInvokeCallback(typeof(Action<IntPtr, string, string>))]
-    static void PopulateDeviceData(IntPtr instance, string key, string value)
-    {
-      var handle = GCHandle.FromIntPtr(instance);
-      if (handle.Target is Device device)
-      {
-        switch (key) {
-          case "jailbroken":
-            switch (value) {
-              case "0":
-                device.AddToPayload(key, false);
-                break;
-              case "1":
-                device.AddToPayload(key, true);
-                break;
-              default:
-                device.AddToPayload(key, value);
-                break;
-            }
-            break;
-          default:
-            device.AddToPayload(key, value);
-            break;
-        }
-      }
-    }
-
-    [StructLayout(LayoutKind.Sequential)]
-    struct NativeUser
-    {
-      public IntPtr Id;
-    }
-
-    public void PopulateUser(User user)
-    {
-      var nativeUser = new NativeUser();
-
-      Wrapper.PopulateUser(ref nativeUser);
-
-      user.Id = Marshal.PtrToStringAuto(nativeUser.Id);
-    }
-
-    public void SetMetadata(string tab, Dictionary<string, string> unityMetadata)
-    {
-      var index = 0;
-      var metadata = new string[unityMetadata.Count * 2];
-
-      foreach (var data in unityMetadata)
-      {
-        metadata[index] = data.Key;
-        metadata[index + 1] = data.Value;
-        index += 2;
-      }
-
-      Wrapper.AddMetadata(NativeConfiguration, tab, metadata, metadata.Length);
-    }
-
-    public void PopulateMetadata(Metadata metadata)
-    {
-    }
-
-    interface ICocoaWrapper
-    {
-      void StartBugsnagWithConfiguration(IntPtr configuration);
-
-      void AddMetadata(IntPtr configuration, string tab, string[] metadata, int metadataCount);
-
-      void RetrieveDeviceData(IntPtr instance, Action<IntPtr, string, string> populate);
-
-      void RetrieveAppData(IntPtr instance, Action<IntPtr, string, string> populate);
-
-      void PopulateUser(ref NativeUser user);
-    }
-
-    const string StartBugsnagWithConfigurationMethod = "bugsnag_startBugsnagWithConfiguration";
-    const string SetMetadataMethod = "bugsnag_setMetadata";
-    const string RetrieveDeviceDataMethod = "bugsnag_retrieveDeviceData";
-    const string RetrieveAppDataMethod = "bugsnag_retrieveAppData";
-    const string PopulateUserMethod = "bugsnag_populateUser";
-
-    class MacOsWrapper : ICocoaWrapper
-    {
-      void ICocoaWrapper.StartBugsnagWithConfiguration(IntPtr configuration) => StartBugsnagWithConfiguration(configuration);
-      void ICocoaWrapper.AddMetadata(IntPtr configuration, string tab, string[] metadata, int metadataCount) => AddMetadata(configuration, tab, metadata, metadataCount);
-      void ICocoaWrapper.RetrieveDeviceData(IntPtr instance, Action<IntPtr, string, string> populate) => RetrieveDeviceData(instance, populate);
-      void ICocoaWrapper.RetrieveAppData(IntPtr instance, Action<IntPtr, string, string> populate) => RetrieveAppData(instance, populate);
-      void ICocoaWrapper.PopulateUser(ref NativeUser user) => PopulateUser(ref user);
-
-      [DllImport(MacOsImport, EntryPoint = StartBugsnagWithConfigurationMethod)]
-      static extern void StartBugsnagWithConfiguration(IntPtr configuration);
-
-      [DllImport(MacOsImport, EntryPoint = SetMetadataMethod)]
-      static extern void AddMetadata(IntPtr configuration, string tab, string[] metadata, int metadataCount);
-
-      [DllImport(MacOsImport, EntryPoint = RetrieveDeviceDataMethod)]
-      static extern void RetrieveDeviceData(IntPtr instance, Action<IntPtr, string, string> populate);
-
-      [DllImport(MacOsImport, EntryPoint = RetrieveAppDataMethod)]
-      static extern void RetrieveAppData(IntPtr instance, Action<IntPtr, string, string> populate);
-
-      [DllImport(MacOsImport, EntryPoint = PopulateUserMethod)]
-      static extern void PopulateUser(ref NativeUser user);
-    }
-
-    class iOSWrapper : ICocoaWrapper
-    {
-      void ICocoaWrapper.StartBugsnagWithConfiguration(IntPtr configuration) => StartBugsnagWithConfiguration(configuration);
-      void ICocoaWrapper.AddMetadata(IntPtr configuration, string tab, string[] metadata, int metadataCount) => AddMetadata(configuration, tab, metadata, metadataCount);
-      void ICocoaWrapper.RetrieveDeviceData(IntPtr instance, Action<IntPtr, string, string> populate) => RetrieveDeviceData(instance, populate);
-      void ICocoaWrapper.RetrieveAppData(IntPtr instance, Action<IntPtr, string, string> populate) => RetrieveAppData(instance, populate);
-      void ICocoaWrapper.PopulateUser(ref NativeUser user) => PopulateUser(ref user);
-
-      [DllImport(iOSImport, EntryPoint = StartBugsnagWithConfigurationMethod)]
-      static extern void StartBugsnagWithConfiguration(IntPtr configuration);
-
-      [DllImport(iOSImport, EntryPoint = StartBugsnagWithConfigurationMethod)]
-      static extern void AddMetadata(IntPtr configuration, string tab, string[] metadata, int metadataCount);
-
-      [DllImport(iOSImport, EntryPoint = RetrieveAppDataMethod)]
-      static extern void RetrieveAppData(IntPtr instance, Action<IntPtr, string, string> populate);
-
-      [DllImport(iOSImport, EntryPoint = RetrieveDeviceDataMethod)]
-      static extern void RetrieveDeviceData(IntPtr instance, Action<IntPtr, string, string> populate);
-
-      [DllImport(iOSImport, EntryPoint = PopulateUserMethod)]
-      static extern void PopulateUser(ref NativeUser user);
-    }
-  }
-}
