+++ src/BugsnagUnity/IBreadcrumbs.cs
