+++ src/BugsnagUnity/IClient.cs
