+++ src/BugsnagUnity/INativeClient.cs
