+++ src/BugsnagUnity/Native/Android/AndroidBreadcrumbs.cs
