+++ src/BugsnagUnity/Native/Android/AndroidClient.cs
