+++ src/BugsnagUnity/Native/Android/AndroidJavaObjectExtensions.cs
