+++ src/BugsnagUnity/Native/Android/DictionaryExtensions.cs
