+++ src/BugsnagUnity/Native/Fallback/Breadcrumbs.cs
