+++ src/BugsnagUnity/Native/Fallback/NativeClient.cs
