+++ src/BugsnagUnity/Native/MacOS/NativeCode.cs
@@ -1,0 +1,7 @@
+﻿namespace BugsnagUnity
+{
+  partial class NativeCode
+  {
+    const string Import = "bugsnag-osx";
+  }
+}
