+++ src/BugsnagUnity/Native/Windows/WindowsClient.cs
