+++ src/BugsnagUnity/Native/iOS/NativeCode.cs
@@ -1,0 +1,7 @@
+﻿namespace BugsnagUnity
+{
+  partial class NativeCode
+  {
+    const string Import = "__Internal";
+  }
+}
